Split AppAuth by provider
